레디스 메모리로 임시변경
